refactor(#34): split cli.ts and register_execute_routes.ts under 300 lines
- cli.ts (350 → 63): subcommand bodies moved to server/src/cli/*.ts
  (print_help, init_config, run_log, run_stats, run_pair, run_server),
  with a small args.ts helper for flag parsing. cli.ts is now a thin
  dispatcher.
- register_execute_routes.ts (353 → 188): the always_approve /
  cached-approval execute-and-audit pattern is hoisted into
  service/execute_and_audit.ts, and the manual-approve try/catch
  (pending-store, abort-on-disconnect, approval race, timeout mapping)
  is hoisted into service/handle_manual_approval.ts.
- Behaviour preserving: audit shape, HTTP status codes, ADR-009 alias
  bypass ordering, and abort-on-disconnect semantics all unchanged.
  All 329 tests pass unchanged.

Closes #34
