+++ server/src/cli/args.ts
@@ -1,0 +1,5 @@
+export function getArgValue(args: string[], flag: string): string | undefined {
+  const idx = args.indexOf(flag);
+  if (idx === -1 || idx >= args.length - 1) return undefined;
+  return args[idx + 1];
+}
